pyproject.toml
[project]
name = "freelancer-data-stack"
version = "0.1.0"
description = "Freelancer Data Stack Project"
authors = [{ name = "Data Team" }]
requires-python = ">=3.11,<3.13"
readme = "README.md"
dependencies = [
    "dbt-core>=1.10.0,<2",
    "dbt-duckdb>=1.9.4,<2",
    "pyan3>=1.2.0",
    "vulture>=2.14",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1,<0.13",
    "black>=25.1.0,<26",
    "isort>=5.13.0,<6",
    "sqlfluff>=3.0.0,<4",
    "pre-commit>=4.0.0,<5",
]
jupyter = [
    "jupyter>=1.1.0,<2",
    "jupyterlab>=4.3.0,<5",
]
datahub = ["acryl-datahub>=0.15.0,<0.16"]
server = [
    "fastapi>=0.115.0,<0.116",
    "uvicorn[standard]>=0.24.0,<0.25",
    "websockets>=12.0.0,<13",
    "strawberry-graphql[fastapi]>=0.275.0,<0.276",
    "snowflake-connector-python[pandas]>=3.15.0,<4",
    "duckdb>=1.3.1,<2",
    "sqlalchemy[asyncio]>=2.0.0,<3",
    "pydantic-ai>=0.0.9,<0.0.10",
    "anthropic>=0.39.0,<0.40",
    "openai>=1.6.0,<2",
    "langchain>=0.3.0,<0.4",
    "langchain-openai>=0.2.0,<0.3",
    "qdrant-client>=1.9.1,<2",
    "pydantic>=2.0.0,<3",
    "python-multipart>=0.0.7,<0.0.8",
    "python-jose[cryptography]>=3.3.0,<4",
    "passlib[bcrypt]>=1.7.0,<2",
    "python-dotenv>=1.0.0,<2",
    "httpx>=0.28.0,<0.29",
    "prometheus-client>=0.19.0,<0.20",
    "structlog>=25.1.0,<26",
]
viz = [
    "streamlit>=1.46.0,<2",
    "pandas>=2.3.0,<3",
    "plotly>=6.2.0,<7",
    "psycopg2-binary>=2.9.0,<3",
]
meltano = [
    "meltano>=3.7.9,<4",
    "singer-sdk>=0.40.0,<0.41",
]

[tool.uv]
package = false
default-groups = [
    "dev",
    "jupyter",
    "datahub",
    "server",
    "viz",
    "meltano",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 88
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.sqlfluff.core]
dialect = "postgres"
max_line_length = 88

[tool.sqlfluff.rules]
max_line_length = 88
